+++ packages/elements/calendar/examples/6-month-and-year.js
@@ -2,7 +2,10 @@
 
 import React from 'react';
 import { CalendarStateless } from '../src';
-import { notThisMonth, notThisYear } from './utils/_';
+
+const thisMonth = new Date().getMonth() + 1;
+const notThisYear = new Date().getFullYear() + 1;
+const notThisMonth = thisMonth === 10 ? 11 : 10;
 
 export default () => (
   <div>