+++ packages/elements/spinner/examples/7-lotsOfSpinners.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import Spinner from '@atlaskit/spinner';
+import Spinner from '../src';
 
 type State = {
   spinners: number[],