+++ packages/elements/toggle/examples/2-disabled.js
@@ -1,6 +1,6 @@
 // @flow
 import React from 'react';
-import Toggle from '../src/';
+import Toggle from '../src';
 
 export default () => (
   <div>