+++ packages/fabric/editor-core/__tests__/ui/ToolbarInsertBlock.tsx
@@ -1,12 +1,8 @@
 import { mount } from 'enzyme';
 import * as React from 'react';
-import blockTypePlugins from '../../src/plugins/block-type';
-import tableCommands from '../../src/plugins/table/commands';
-import DropdownMenu from '../../src/ui/DropdownMenu';
-import ToolbarInsertBlock from '../../src/ui/ToolbarInsertBlock';
-import { EmojiPicker as AkEmojiPicker } from '@atlaskit/emoji';
-import { testData as emojiTestData } from '@atlaskit/emoji/dist/es5/support';
 import Item from '@atlaskit/item';
+import { EmojiPicker as AkEmojiPicker, EmojiProvider } from '@atlaskit/emoji';
+import { testData as emojiTestData } from '@atlaskit/emoji/dist/es5/support';
 import {
   doc,
   p,
@@ -14,14 +10,21 @@ import {
   code_block,
   defaultSchema,
 } from '@atlaskit/editor-test-helpers';
-import ToolbarButton from '../../src/ui/ToolbarButton';
-import { MediaProvider } from '@atlaskit/media-core';
 import { ProviderFactory } from '@atlaskit/editor-common';
+
+import blockTypePlugins from '../../src/plugins/block-type';
+import tableCommands from '../../src/plugins/table/commands';
+import DropdownMenu from '../../src/ui/DropdownMenu';
+import ToolbarInsertBlock from '../../src/ui/ToolbarInsertBlock';
+import ToolbarButton from '../../src/ui/ToolbarButton';
 import codeBlockPlugin from '../../src/editor/plugins/code-block';
 import panelPlugin from '../../src/editor/plugins/panel';
 import listPlugin from '../../src/editor/plugins/lists';
+import { MediaProvider } from '../../src/plugins/media';
 
-const emojiProvider = emojiTestData.getEmojiResourcePromise();
+const emojiProvider = emojiTestData.getEmojiResourcePromise() as Promise<
+  EmojiProvider
+>;
 
 const mediaProvider: Promise<MediaProvider> = Promise.resolve({
   viewContext: Promise.resolve({} as any),
@@ -339,6 +342,30 @@ describe('@atlaskit/editor-core/ui/ToolbarInsertBlock', () => {
     expect(insertMacroFromMacroBrowser).toHaveBeenCalledTimes(1);
     expect(trackEvent).toHaveBeenCalledWith(
       'atlassian.editor.format.macro.button',
+    );
+    toolbarOption.unmount();
+  });
+
+  it('should track placeholder insert event when "Add Placeholder Text" option is clicked', () => {
+    const { editorView } = editor(doc(p('text')));
+
+    const toolbarOption = mount(
+      <ToolbarInsertBlock
+        placeholderTextEnabled={true}
+        editorView={editorView}
+        buttons={0}
+        isReducedSpacing={false}
+      />,
+    );
+
+    toolbarOption.find(ToolbarButton).simulate('click');
+    const button = toolbarOption
+      .find(Item)
+      .filterWhere(n => n.text().indexOf('Placeholder Text') > -1);
+    button.simulate('click');
+
+    expect(trackEvent).toHaveBeenCalledWith(
+      'atlassian.editor.format.placeholder.button',
     );
     toolbarOption.unmount();
   });