+++ packages/fabric/editor-core/src/nodeviews/ui/mention.tsx
@@ -15,14 +15,13 @@ export interface Props {
 export default class MentionNode extends PureComponent<Props, {}> {
   render() {
     const { node, providerFactory } = this.props;
-    const { id, text, accessLevel, userType } = node.attrs;
+    const { id, text, accessLevel } = node.attrs;
 
     return (
       <Mention
         id={id}
         text={text}
         accessLevel={accessLevel}
-        userType={userType}
         providers={providerFactory}
       />
     );