+++ packages/fabric/media-card/src/shared/CardFrame/styled.ts
@@ -10,6 +10,7 @@ import {
 import { borderRadius, size, linkCardShadow, ellipsis } from '../../styles';
 
 const wrapperPadding = 8;
+export const className = 'media-card-frame';
 
 export interface WrapperProps {
   isInteractive?: boolean;
@@ -133,12 +134,26 @@ export const TextWrapper = styled.div`
   ${ellipsis('none')};
 `;
 
+export interface ContentProps {
+  isInteractive: boolean;
+}
+
 export const Content = styled.div`
+  position: relative;
+
   ${borderRadius} ${linkCardShadow} background-color: white;
   transition: box-shadow 0.3s;
 
-  .media-card-link:hover & {
-    box-shadow: 0 4px 8px -2px rgba(23, 43, 77, 0.32),
-      0 0 1px rgba(23, 43, 77, 0.25);
-  }
+  ${({ isInteractive }: ContentProps) => {
+    if (isInteractive) {
+      return ` 
+          .${className}:hover & {
+            box-shadow: 0 4px 8px -2px rgba(23, 43, 77, 0.32),
+              0 0 1px rgba(23, 43, 77, 0.25);
+          }
+        `;
+    } else {
+      return '';
+    }
+  }};
 `;