+++ packages/elements/analytics/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/analytics
 
+## 2.4.2
+- [patch] Packages Flow types for elements components [3111e74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3111e74)
+
 ## 2.4.1
 - [patch] Fix clean props type [b0a2cf0](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b0a2cf0)
 - [patch] Add a typescript definition for clean props [88cc4be](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/88cc4be)