+++ packages/elements/avatar/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/avatar
 
+## 9.0.1
+- [patch] Packages Flow types for elements components [3111e74](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/3111e74)
+
 ## 9.0.0