+++ packages/fabric/media-core/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/media-core
 
+## 14.0.0
+- [major] Move media provider and state manager to editor-core [0601da7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0601da7)
+
 ## 13.0.0